Split ECSqlFuzzGenerator.h into 4 files for maintainability
Split the 3284-line monolithic generator into focused modules:
- ECSqlFuzzGenerator.h (525 lines): Core class, helpers, GenAnyStatement dispatcher
- ECSqlFuzzSelectGenerators.h (2008 lines): All SELECT query generators
- ECSqlFuzzWriteGenerators.h (455 lines): INSERT/UPDATE/DELETE/PRAGMA/schema generators
- ECSqlFuzzMutationEngine.h (318 lines): Mutation engine, boundary inputs, deep nesting

Sub-files are included inside the struct body via #include.
